Split Log4j1 config factory into tiny factory and separate parser
